fix(docker): Copy config files into Dockerfile instead of mounting them as a volume
This is a fix related to remote Docker host which can bind mount a file or directory.
Bind mounts is not a requirement for us, this should be a safe workaround.

It may need some changes in the workshop documentation: explain where are files and how to redeploy a component after config change.